articleLike, productLike로 코드 분리
