+++ src/services/likeService.ts
@@ -1,7 +1,8 @@
-import likeRepository from "../repositories/likeRepository";
 import userReporitory from "../repositories/userReporitory";
 import productRepository from "../repositories/productRepository";
 import articleReporitory from "../repositories/articleReporitory";
+import productLikeRepository from "../repositories/productLikeRepository";
+import articleLikeReporitory from "../repositories/articleLikeReporitory";
 
 export class LikeService {
     async updateProductLike(userId: number, productId: number) {
@@ -15,19 +16,19 @@ export class LikeService {
             throw new Error('Product not found')
         };
 
-        const existingLike = await likeRepository.checkingProductLikeStatus(userId, productId);
+        const existingLike = await productLikeRepository.checkingProductLikeStatus(userId, productId);
         let message: string;
         let isLiked: boolean;
         let likeCount: number;
 
         if (existingLike) {
-            await likeRepository.deleteProductLike(existingLike.id);
+            await productLikeRepository.deleteProductLike(existingLike.id);
             const updatedProduct = await productRepository.updateLikeDecrease(productId);
             message = 'Product unliked successfully'
             isLiked = false;
             likeCount = updatedProduct.likeCount;
         } else {
-            await likeRepository.uploadProductLike(userId, productId);
+            await productLikeRepository.uploadProductLike(userId, productId);
             const updatedProduct = await productRepository.updateLikeIncrease(productId);
             message = 'Product liked successfully';
             isLiked = true;
@@ -47,19 +48,19 @@ export class LikeService {
             throw new Error('Article not found');
         }
 
-        const existingLike = await likeRepository.checkingArticleLikeStatus(userId, articleId);
+        const existingLike = await articleLikeReporitory.checkingArticleLikeStatus(userId, articleId);
         let message: string;
         let isLiked: boolean;
         let likeCount: number;
 
         if (existingLike) {
-            await likeRepository.deleteArticleLike(existingLike.id);
+            await articleLikeReporitory.deleteArticleLike(existingLike.id);
             const updatedArticle = await articleReporitory.updateLikeDecrease(articleId);
             message = 'Article unliked successfully';
             isLiked = false;
             likeCount = updatedArticle.likeCount;
         } else {
-            await likeRepository.uploadArticleLike(userId, articleId);
+            await articleLikeReporitory.uploadArticleLike(userId, articleId);
             const updatedArticle = await articleReporitory.updateLikeIncrease(articleId);
             message = 'Article liked successfully';
             isLiked = true;
